Refactor, cleanup. Seems to work ok.
